[4.3] SUPP-16: load_parents sets an empty object for account (#6627)
It is not clear what is the difference between `/accounts/:id/parents`
and `/accounts/:id/tree`. But the are doing same thing in different
ways.

As it turns out `load_parents` is loading all result of
`accounts/listing_by_id` and then uses `kz_json:find_value/3` to grab
the parent's account info.

But it is constantly failing in a specific server. I can't reproduce
the exact problem in any environment that I test it.

For example it should return:

```json
[{"id":"{ACCOUNT_1}","name":"{ACCOUNT1_NAME}"},{"id":"{ACCOUNT_2}","name":"{ACCOUNT2_NAME}"}]
```

But it returns instead:

```json
[{"id":"{ACCOUNT_1}","name":"{ACCOUNT1_NAME}"},{}]
```

Since those `/tree` and `/parents` seems returning same thing and
`/tree` implementation is way more efficient and stable than
`/parents` I change it to use that.